+++ Accelerometer/ADXL372/adxl372_spi_nolib.h
